moving to one schema.ts file
