+++ src/db/schema.ts
@@ -350,6 +350,24 @@ export const pubmedDailyUpdate = pgTable('pubmed_daily_update', {
   xml_filename: text('xml_filename'),
 });
 
+// Keycloak user_entity table schema
+export const keycloakUsers = pgTable('user_entity', {
+  id: text('id').primaryKey(),
+  email: text('email'),
+  email_constraint: text('email_constraint'),
+  email_verified: boolean('email_verified').notNull().default(false),
+  enabled: boolean('enabled').notNull().default(false),
+  federation_link: text('federation_link'),
+  first_name: text('first_name'),
+  last_name: text('last_name'),
+  realm_id: text('realm_id'),
+  username: text('username'),
+  created_timestamp: bigint('created_timestamp', { mode: 'number' }),
+  service_account_client_link: text('service_account_client_link'),
+  not_before: integer('not_before').notNull().default(0)
+});
+
+
 // Table relationships
 export const llmGuidedSectionsRelations = relations(llmGuidedSections, ({ many }) => ({
   contexts: many(llmGuidedContexts, { relationName: 'section_contexts' }),
@@ -477,3 +495,7 @@ export type NewProjectStats = typeof projectStats.$inferInsert;
 
 export type PubmedDailyUpdate = typeof pubmedDailyUpdate.$inferSelect;
 export type NewPubmedDailyUpdate = typeof pubmedDailyUpdate.$inferInsert;
+
+// export types for keycloak users
+export type KeycloakUsers = typeof keycloakUsers.$inferSelect;
+export type NewKeycloakUsers = typeof keycloakUsers.$inferInsert;
